feat(function): centralized webhooks for plugins
Adds cloud function located at `/webhook`, which plugins can register their webhooks in the database as `webhooks/tokens/{webhooktoken}/{CID}/` (`{cidhook}`).

Presence is registered as `{cidhook}/{key}=true`, to tell the core that the webhook exists. If a webhook is not found at the time the hook is received the server responds with `Error Code 400: Bad Request`.

The hook data is then passed along to `webhooks/data/{KEY}/` (`{keydata}`) and plugins listen for new children at `{cidhook}/keys/`, where the key will be added as `{cidhook}/keys/{KEY}=true`.

At this point plugins are expected to read `{keydata}` and set `{keydata}/status=statuscode`.

Core waits up to 5 seconds for a status code before returning 504 (gateway timeout), or the status code which is set. All data and keys are deleted from the database after core has responded.